requirements.txt
discord.py>=2.0.0

pydantic>=1.9.2
python-dotenv>=0.20.0

# formatting
black>=22.6.0
